add sorting intercept
